move AgentWithChatRoom to chatroom.py
