+++ src/agentscope/manager/_monitor.py
@@ -55,11 +55,10 @@ class MonitorManager:
                 cls,
             )
         else:
-            pass
-            # raise RuntimeError(
-            #     "The monitor manager has been initialized. Try to use "
-            #     "MonitorManager.get_instance() to get the instance.",
-            # )
+            raise RuntimeError(
+                "The monitor manager has been initialized. Try to use "
+                "MonitorManager.get_instance() to get the instance.",
+            )
         return cls._instance
 
     @property
